+++ changeLog.md
@@ -18,3 +18,8 @@ Note that if you want to search against a property name you have now to put in u
 
 * hide pagination when less than 1 page
 * add unit tests for pagination directive
+
+## version 1.2.3
+
+* fix back to natural sort order
+* use same strategy view -> table state, table state -> view for all the plugins